Don't create annotations if not creating the service account
Currently setting the serviceAccountAnnotations without the other rbac stuff fails with `error validating data: [apiVersion not set, kind not set]`.
This change would prevent failure, but will swallow a potential misconfiguration (would a user want to have serviceAccountAnnotations with the service account not being created? - maybe but I'm not too sure about this).
As an alternative, an explicit fail message could be set if serviceAccountAnnotations are set without the service account being created.